Split up DB retrieving and adding code
