+++ app/src/main/java/de/upb/hip/mobile/activities/DisplayImageSliderActivity.java
@@ -30,8 +30,8 @@ import java.util.List;
 
 import de.upb.hip.mobile.adapters.DBAdapter;
 import de.upb.hip.mobile.helpers.CustomSeekBar;
+import de.upb.hip.mobile.helpers.db.ExhibitDeserializer;
 import de.upb.hip.mobile.models.exhibit.Exhibit;
-import de.upb.hip.mobile.models.exhibit.ExhibitDeserializer;
 
 /**
  * Activity Class for the Image-Slider View, where a picture at the top can be changed with a fading
